requirements.txt
google-cloud-bigquery == 2.34.4
google-cloud-storage == 2.8.0
fastapi == 0.110.0
pydantic == 2.7.0
uvicorn == 0.17.6
jinja2 == 3.1.0
python-jose == 3.3.0
python-multipart==0.0.5
passlib[bcrypt] == 1.7.4